chore(parser): move get operation name to common parser
