+++ src/openApi/common/parser/__tests__/operation.spec.ts
