+++ src/openApi/v2/parser/getOperation.ts
@@ -1,6 +1,6 @@
 import type { Operation, OperationParameters } from '../../../types/client';
 import type { Config } from '../../../types/config';
-import { getOperationName } from '../../../utils/operation';
+import { getOperationName } from '../../common/parser/operation';
 import { toSortedByRequired } from '../../common/parser/sort';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiOperation } from '../interfaces/OpenApiOperation';
